refactor(transport): extract endpoint arbitration; present-tense comments
ZmqTransport carried ~130 lines of POSIX endpoint arbitration — flock
ownership, an AF_UNIX connect(2) liveness probe, and their rationale —
inline in the class a learner reads to understand 'how does the C++ side
talk to the emulator'. That machinery now lives in its own translation
unit (endpoint_lock.hpp/.cpp) with its prose intact, and ServerThread's
bind phase reads as the three lines of intent it always was: take the
lock, probe for a foreign owner, bind.

TransportConfig is an aggregate again. Its Logger& member forced two
user-provided constructors, which killed designated initializers — the
idiom the rest of the codebase leans on — and both transport test files
had grown a MakeConfig helper with an apologetic comment to work around
it. The member is now a reference_wrapper defaulting to DefaultLogger(),
callers write TransportConfig{.send_timeout = ..., .logger = ...}, and
the workarounds are gone.

Comments that narrated the code's history in past tense are rewritten to
describe present behavior; the history lives here in the log instead.
The 'reduce cognitive complexity' apology over the Log helpers is gone —
they stay on their merits as brevity.

Also: delay.hpp puts #pragma once first, names its parameter duration
(callers pass milliseconds; only the conversion target was micro), and
documents the blocking semantics; transport.hpp loses its empty private
section and gains [[nodiscard]] and an interface comment.
