+++ src/libs/mcu/delay.hpp
@@ -1,8 +1,13 @@
-#include <chrono>
-
 #pragma once
 
+#include <chrono>
+
 namespace mcu {
-auto Delay(std::chrono::microseconds usecs) -> void;
+
+/// @brief Block the calling thread for at least `duration`.
+/// Callers pass any chrono duration (e.g. 200ms); it converts implicitly.
+/// On the host this is sleep_for, so only the calling thread pauses — the
+/// transport's server thread keeps handling emulator messages throughout.
+auto Delay(std::chrono::microseconds duration) -> void;
 
 }  // namespace mcu
